This is an initial commit for the multi-agent project.
At this stage, the value network involves an RNN to
process sequences of frames.  Deep Q-learning is implemented
by the agent (with DeepMind's experience trick).  During
learning recall phases, rather than remembering entire
sequences of frames, the agent recalls hidden LSTM states
that should offer an approximation to the effect
imputting the full sequence of states.

Unfortunately, testing so far has shown that this
approach does a poor job of learning to remember things.
The agent is bad at tracking where it has already looked.
The problem is probably connected to the use of
Q-learning rather than a policy-gradient method.